+++ src/incident_reports/incident_reports_server/factory/incident_report_factory.py
@@ -1,5 +1,5 @@
 import json
-from incident_reports_server.model.models import IncidentReport, Severity, Status
+from incident_reports_server.models.models import IncidentReport, Severity, Status
 
 class IncidentReportFactory:
