+++ src/incident_reports/incident_reports_server/persistence/i_incident_report_persistence.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import List
 
-from incident_reports_server.model.models import IncidentReport
+from incident_reports_server.models.models import IncidentReport
 
 class IIncidentReportPersistence(ABC):
     @abstractmethod
